# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Docmaster API",
        "description": "Docmaster API Documentation",
        "contact": {
            "name": "LJN",
            "email": "jordanndjock07@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/document": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Liste des documents de l'utilisateur authentifié",
                "description": "Nécessite un token Bearer dans l'en-tête Authorization",
                "operationId": "17924497db843e276fc8a03cca6d6574",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "user_documents": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Document"
                                                    }
                                                },
                                                "meta": {
                                                    "type": "object",
                                                    "example": {
                                                        "total": 20,
                                                        "current_page": 1,
                                                        "per_page": 10
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des documents de l'utilisateur récupérée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Créer un document",
                "description": "Nécessite un token Bearer dans l'en-tête Authorization",
                "operationId": "85804a39a2c423b4aae96724d2b0ba89",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "type_document_id",
                                    "nom_proprietaire",
                                    "titre",
                                    "fichier_document"
                                ],
                                "properties": {
                                    "type_document_id": {
                                        "type": "string",
                                        "example": "uuid-type-doc"
                                    },
                                    "nom_proprietaire": {
                                        "type": "string",
                                        "example": "Jean Michel"
                                    },
                                    "titre": {
                                        "type": "string",
                                        "example": "Permis de conduire"
                                    },
                                    "date_expiration": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    },
                                    "fichier_document": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Document créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Document"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "document créé avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/document/{id}": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Afficher un document",
                "description": "Nécessite un token Bearer dans l'en-tête Authorization",
                "operationId": "a64b467a6bdacc7af4279c936970a1f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document récupéré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Document"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Document récupéré avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Document non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Documents"
                ],
                "summary": "Mettre à jour un document",
                "description": "Nécessite un token Bearer dans l'en-tête Authorization",
                "operationId": "d60a177265ceb59f9d0fabbb4d2b457c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "type_document_id",
                                    "nom_proprietaire",
                                    "titre",
                                    "fichier_document"
                                ],
                                "properties": {
                                    "type_document_id": {
                                        "type": "string",
                                        "example": "uuid-type-doc"
                                    },
                                    "nom_proprietaire": {
                                        "type": "string",
                                        "example": "Jean Michel"
                                    },
                                    "titre": {
                                        "type": "string",
                                        "example": "Permis de conduire"
                                    },
                                    "date_expiration": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    },
                                    "fichier_document": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Document modifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Document"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "document modifié avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Documents"
                ],
                "summary": "Archiver un document",
                "description": "Nécessite un token Bearer dans l'en-tête Authorization",
                "operationId": "b524df9bd74a189b47c69e0b4112c006",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document archivé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Document archivé avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/document/{id}/restore": {
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Restaurer un document",
                "description": "Nécessite un token Bearer dans l'en-tête Authorization",
                "operationId": "1b1820cac6fcfb9b37fac7f5ced8cd60",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document restauré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Document restauré avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/document/{id}/force-delete": {
            "delete": {
                "tags": [
                    "Documents"
                ],
                "summary": "Supprimer définitivement un document",
                "description": "Nécessite un token Bearer dans l'en-tête Authorization",
                "operationId": "2f8a2227cd82d9c6093af31961fed315",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document supprimé définitivement",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Document supprimé définitivement avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Document": {
                "required": [
                    "id",
                    "user_id",
                    "type_document_id",
                    "nom_proprietaire",
                    "titre",
                    "fichier_url"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "uuid-1234"
                    },
                    "user_id": {
                        "type": "string",
                        "example": "uuid-user-1"
                    },
                    "type_document_id": {
                        "type": "string",
                        "example": "uuid-type-doc"
                    },
                    "nom_proprietaire": {
                        "type": "string",
                        "example": "Jean Michel"
                    },
                    "titre": {
                        "type": "string",
                        "example": "Acte de naissance"
                    },
                    "date_expiration": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-12-31"
                    },
                    "fichier_url": {
                        "type": "string",
                        "example": "documents/acte-naissance.pdf"
                    },
                    "trouve": {
                        "type": "boolean",
                        "example": false
                    },
                    "sauvegarde": {
                        "type": "boolean",
                        "example": false
                    },
                    "signale": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}